wip: app_config settings and neatening
