--- src/containers/Settings/MiningSettings.tsx
@@ -1,13 +1,0 @@
-import CpuMiningMarkup from './sections/mining/CpuMiningMarkup.tsx';
-import GpuMiningMarkup from './sections/mining/GpuMiningMarkup.tsx';
-import MineOnStartMarkup from './sections/mining/MineOnStartMarkup.tsx';
-
-export const MiningSettings = () => {
-    return (
-        <>
-            <CpuMiningMarkup />
-            <GpuMiningMarkup />
-            <MineOnStartMarkup />
-        </>
-    );
-};
