+++ src/containers/Settings/sections/experimental/Temp.tsx
@@ -1,0 +1,9 @@
+import PaperWallet from '@app/containers/Settings/sections/wallet/PaperWallet.tsx';
+
+export const TempSettings = () => {
+    return (
+        <>
+            <PaperWallet />
+        </>
+    );
+};
