+++ src/containers/Settings/sections/mining/MiningSettings.tsx
@@ -1,0 +1,13 @@
+import CpuMiningMarkup from './CpuMiningMarkup.tsx';
+import GpuMiningMarkup from './GpuMiningMarkup.tsx';
+import MineOnStartMarkup from './MineOnStartMarkup.tsx';
+
+export const MiningSettings = () => {
+    return (
+        <>
+            <CpuMiningMarkup />
+            <GpuMiningMarkup />
+            <MineOnStartMarkup />
+        </>
+    );
+};
